app/server/requirements/local.txt
-r base.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5; python_version >= '3.6'
attrs==23.2.0; python_version >= '3.7'
coverage[toml]==7.6.0; python_version >= '3.8'
flake8==7.1.0; python_full_version >= '3.8.1'
iniconfig==2.0.0; python_version >= '3.7'
mccabe==0.7.0; python_version >= '3.6'
packaging==24.1; python_version >= '3.8'
pluggy==1.5.0; python_version >= '3.8'
psycopg==3.2.1; python_version >= '3.8'
pycodestyle==2.12.0; python_version >= '3.8'
pyflakes==3.2.0; python_version >= '3.8'
pytest==8.3.2; python_version >= '3.8'
pytest-cov==5.0.0; python_version >= '3.8'
pytest-django==4.8.0; python_version >= '3.8'
pytest-sugar==1.0.0
termcolor==2.4.0; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
